Consolidate localization and UI refresher implementations into a single file; remove obsolete classes
