+++ src/Bucket.Core/Bucket.Core.csproj
@@ -1,8 +1,16 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
-    <TargetFramework>net9.0</TargetFramework>
+    <TargetFramework>net9.0-windows10.0.26100</TargetFramework>
+    <TargetPlatformMinVersion>10.0.17763.0</TargetPlatformMinVersion>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
+    <UseWinUI>true</UseWinUI>
   </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="DevWinUI" Version="9.1.0" />
+    <PackageReference Include="WinUI3Localizer" Version="2.3.0" />
+    <PackageReference Include="Microsoft.WindowsAppSDK" Version="1.8.250907003" />
+  </ItemGroup>
 </Project>
